Create EXIF data embedding script
location_embed.py is a tool to embed location data in satellite images. Reads a list of filenames and location data from a csv, and then searches root and subdirectories for matching files. Matched files have their exif data modified. A log of matched/unmatched files is also written to disk.
Needs further modifications and is so far untested.