New helper functions for location embedding, additional misc fixes
New function to perform coordinate conversion to/from decimal degrees and DMS formats. New function to update location data associated with each image, before it's written to the EXIF fields. Additional fixes made to the file_enum function to fix data type error, and to the exifwrite function to reflect changes in the passed arguments