Bugfixes, more problems
The embedding operation finally seems to somewhat work, however the values don't show up when viewing the image description.
I believe that the EXIF structure in the file becomes non-standard/non-compliant with the specification when I try to insert the tags. This is also because there is a GPS pointer tag which is meant to point to the start of the GPS information segment. My course of action from now on will be to formulate the basic EXIF structure from scratch instead of attempting to modify the existing structure within the image.